Fix the fuzzing of the DFFs.
The situation where an extra bit was captured meant that
only two identical registers could be placed in a single slide.
It is now possible to place registers of joint types, i.e:
DFFS and DFFR
DFFP and DFFC
DFFNS and DFFNR
DFFNP and DFFNC
as well as their CE versions.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>