Place logicinfo in the database.
The logicinfo table describes the valid values of the parameters of the
primitives (bels). Each line is a pair (parameter_id, value).
This table can be used to check if the parameter values are set
correctly, but its most important purpose is to generate fuses that
program the above parameter value.  The fact is that the first two
fields of the shortval table are the row numbers of this logicinfo
table, that is, the fuses that are listed in the row of the shortval
table must be set to set the values of this parameter pair.

As long as there is no decoding of parameter_id it is hard to give a
clear example.  From unconfirmed theory in the longval table, the first
six fields also refer to logicinfo.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>