Add longval table processing
This function is the heart of this table system: chipdb.get_longval_fuses()

It takes as input a set of primitive attribute values and returns bits to set.
Example usage:
ipdb> chipdb.get_longval_fuses(db, 66, {19, 56, 72}, 'IOBB')
(True, {(3, 53), (3, 43), (2, 46), (3, 41), (3, 42), (3, 39)})

Of course, we need more partner functions to work with attributes, and
we need symbolic names for attributes and values.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>