Add a mechanism to convert attribute/value pairs
The add_attr_val() function adds the index of a valid attribute and
value combination to the set.  If the combination is invalid, it is
silently ignored. This is a consequence of the fact that the "logicinfo"
table is incomplete: it lacks quite good combinations, which are set by
default.
But since they (the combinations) are already set by default, this
mechanism is fine for coding, but it is not suitable for checking for
permissibility.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>