Remove all attribute setting checks
For really big primitives, those that occupy more than one grid cell,
it's much more convenient to pass all the attributes of the primitive in
a pile without having to deal with which particular cell implements
what. Especially since this can change from chip to chip.
So now the fuse get function no longer checks for extra/erroneous
attributes, it returns bits for what it finds.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>