Add the alonenode mechanism (#150)
Storing and using part of the "alonenode" table to set fuses designed to
disable "lone" nodes, i.e. nodes that are not used as sources. It is
used to switch long wires in buses.

In addition, the representation of long wires is greatly simplified by
creating only the necessary aliases and transferring the main work to
the nextpnr.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>