stage 2. Add GW1NSR-4C PLLs
A new type of PLL is added - PLLVR, which is used in the GW1NS-4 chips.
Both PLL instances are packaged and unpacked with support for all
attributes, including static and dynamic setting of generated frequency
parameters.

Describes the implicit wires leading to the central clock MUX, which
allows us to use all 4 PLL outputs using the global clock networks.

Also the meaning of Table 59 and a couple of attributes in it was
discovered. Not used yet, the constants added for the future.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>