Add support for OSER primitives
Primitives are implemented that convert the parallel signals into a
serial signal. The primitives differ in the number of input parallel
signals: 4, 7, 8 and 10 respectively for OSER4, OVIDEO, OSER8 and OSER10
primitives.

The serialization primitives require two clock frequencies: one slow
(pclk) is used to load parallel data, and the other fast (fclk) is used
to generate the serial output signal.

For pclk, normal CLK-type ports are used, which under certain conditions
can use routing through global clock networks or ordinary wires -
nothing unusual here.

But fclk uses a special kind of wires - HCLK, of which there are very
few: 4 wires per side of the chip, and you have to choose in each
particular IO buffer from only two wires.

Not all chips have all sides covered by HCLK: the GW1N-1 has only the
bottom side, the GW1NZ-1 has only the top and right side, the GW1NSR-4C
has all but the left side, etc.

The situation is not improved by the fact that these wires are
controlled in special cells, the nature and number of which depend on
the chip.

This commit implements the HCLK capabilities necessary for the operation
of the serialization primitives.

The OSER4, OVIDEO, OSER8 and OSER10 primitives are implemented for the
GW1N-1, GW1NZ-1, GW1N-4, GW1NSR-4C, GW1NR-9 and GW1NR-9C chips.

Demo examples are added for all primitives. On the tangnano9k board you
can observe the operation immediately on the LEDs (because there are 6
of them), for other boards you will need a logic analyzer.
Signed-off-by: YRabbit <rabbit@yrabbit.cyou>