Add support for IDES primitives
Primitives are implemented that convert the serial signal into a
parallel signals. The primitives differ in the number of input parallel
signals: 4, 7, 8 and 10 respectively for IDES4, IVIDEO, IDES8 and IDES10
primitives.

The primitives are implemented for the GW1N-1, GW1NZ-1, GW1N-4, GW1NSR-4C,
GW1NR-9 and GW1NR-9C chips.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>