Adapting the packer to the alternative P&R
nextpnr with gowin as the Himbaechel architecture mainly affects gowin_pack.

At this experimental stage, I think it is wise not to mimic the old
input format, but to support both for now.

The new nextpnr output file has a format flag in the header, so it's
easy to distinguish between the two.

- new cell numbering, new wire naming, pips and bels are used.
- LUT and DFF come separately, not as SLICE - I'm interested to see what
  comes out of this and how it affects all sorts of ALU and RAM
- IO is purely nominal for now, just to make the LEDs and buttons work
  during the experiments.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>