Add clock lines for Himbaechel router
Clock wires are described as Himbaechel nodes, which greatly reduces the
code on the P&R side - there is no need for additional processing as for
aliases.

The sources for clocks that are dedicated pins are extracted and described.

The clock fuzzer stays in place until there is a complete transition to
the Himbaechel architecture - describing a clock network using aliases
that describe the permanent connection of only a pair of wires quickly
ceases to reflect the reality of having multiple wires permanently
connected together.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>