Himbaechel-gowin. Add the examples.
For each supported board (GW1NS-2 excluded), the following examples are provided:
- blinky,
- shift,
- blinky-tbuf,
- blinky-oddr,
- blinky-osc,
- blinky-pll*,
- elvds,
- oddr-elvds,
- tlvds,
- oddr-tlvds,
- pll-nanolcd*.
* - where there is (is not available) a VGA connector.
Attosoc is also provided for larger boards.

For small boards with a VGA connector, the -noalu flag is used, but only
because the required number of primitives does not fit into the board.

There are no examples of DESER yet, not because of the functionality of the primitives.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>