Himbaechel. Improve dedicated clock router
Added buffers for the global clock network.
We indicate that the network is CLOCK by placing the following
construction in the file of restrictions:

CLOCK_LOC "net_name" BUFG;

As a result, the specified network is cut into two parts: the part from
the source to the buffer is routed using any available PIPs, while the
part from the buffer to the sink is routed through the global clock
network.

This will slightly improve the situation with Tangnano20k, since now it
is possible to declare its pin, on which the simple external generator
is located, as a clock, which will lead to routing through the buffer
without any clock glitches.

All examples for Tangnano20k, which previously used the PLL as a buffer,
are no longer needed and use the same files as for the rest of the
boards.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>