Himbaechel. Add BSRAM for all chips.
The following primitives are implemented for the GW1N-1, GW2A-18,
GW2AR-18C, GW1NSR-4C, GW1NR-9C, GW1NR-9 and GW1N-4 chips:

    * pROM     - read only memory - (bitwidth: 1, 2, 4, 8, 16, 32).
    * pROMX9   - read only memory - (bitwidth: 9, 18, 36).
    * SDPB     - semidual port    - (bitwidth: 1, 2, 4, 8, 16, 32).
    * SDPX9B   - semidual port    - (bitwidth: 9, 18, 36).
    * DPB      - dual port        - (bitwidth: 16).
    * DPX9B    - dual port        - (bitwidth: 18).
    * SP       - single port      - (bitwidth: 1, 2, 4, 8, 16, 32).
    * SPX9     - single port      - (bitwidth: 9, 18, 36).

For GW1NSR-4C and GW1NR-9 chips, SP/SPX9 primitives with data widths of
32/36 bits are implemented using a pair of 16-bit wide primitives.

Added examples for all boards except those based on GW1NSR-4C, GW1NR-9
and GW1N-4 chips for memory primitives with a width of 8 bits (as well
as 16 bits where 8 is not supported).

And these very examples are the weakest point - I tried to make the
primitives themselves work and the result of compiling the examples was
similar between the vendor IDE and Apicula, but the examples themselves
were poorly written. This is due to my lack of experience working with
BSRAM.  In general, they need to be rewritten.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>