Add support for four IOLOGICs in one cell
There's actually very little going on on the apicula side, but a future
nextpnr fix will split each of IOLOGICA and IOLOGICB into two
directions: IOLOGIC in and IOLOGIC out. This allows us to combine two
IOLOGICs at once to work with IOBUF, which addresses issue
https://github.com/YosysHQ/nextpnr/issues/1275

The nature of the fixes in this PR is such that it will work with the
old version of nextpnr and everything that successfully passes through
netxpnr will work.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>