Implement the DSP primitive.
For chips that have these capabilities, a DSP implementation has been
added in the form of all the primitives described in the Gowin
documentation (UG287-1.3.3E_Gowin Digital Signal Processing (DSP) User
Guide), namely:

  - PADD9
  - PADD18
  - MULT9X9
  - MULT18X18
  - MULT36X36
  - MULTALU18X18
  - MULTALU36X18
  - MULTADDALU18X18
  - ALU54D

The most complex but also the most useful is the MULTADDALU18X18
primitive - it allows you to easily make a typical FIR filter, while all
connections between these primitives in the chain will be implemented by
direct fixed wires with minimal delay.

MULT36X36 are not combined into chains, but they have a different task -
this primitive can be found in Linux SOCs.

Added examples (in the examples/himbaechel directory) that are based on
the tiny Riscv demonstrating UART calculations. Only the TXD pin is used
(can be found in the specific .CST file for each board), so on the large
computer side, only GND and RXD are enough. Port speed 115200, no
parity, 8 data bits, 1 stop bit, linefeed only.

Picocom launch example:

``` shell
picocom -l --imap lfcrlf -b 115200 /dev/ttyU0
```

The source code for the riscv test programs is provided along with the
assembly instructions, but they are not built during the compilation of
the examples due to additional compilers.

Implemented the combination of primitives into chains using wires
CASO-CASI, SO(A, B)-SI(A, B), as well as SBO-SBI for PADD.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>