Add PLL pads.
Whenever possible, we arrange the PLL so as to use a direct connection
of pins to the inputs of the primitive. The main work happens in
nextpnr-himbaechel, here we only prepare data about which pin
corresponds to which PLL.

Also added an example for Tangnano4k where two PLLs compete for one
input pin. Only one of them can win, but both must function.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>