Add dynamic clock control
DQCE and DCS primitives are added - disabling the clock in a quadrant
and dynamically selecting the clock source accordingly.

With this commit we are finally able to use clock lines 6 and 7 :)

Added examples for Tangnano1k, Tangnano4k, Tangnano9k, Primer20k,
Tangnano20k boards.

From the internals: in addition to the well-known fuse tables SHORTVAL
and LONGVAL, functions have been added for working with another table -
LONGFUSE. It differs in that it describes the fuses of primitives with
only one parameter, in contrast to two parameters for SHORTVAL and 16
parameters for LONGVAL.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>