HCLK. Add interbank wires
An interesting mechanism for transmitting signals between HCLK banks (or
otherwise parties) has been discovered.

We have two points, each of which can receive a signal from any HCLK and
transmit it to any other HCLK.

No complex signals can be transmitted here - the input lines are limited
by HCLK_INx, and we also need to see which components can use the output
from these two points, but so far the tests are encouraging: it is
possible to transmit a signal from the clock pin from one side of the
chip to the opposite and feed CLKDIV2 there using just a couple wires!

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>