Implement EMCU primitive
In this PR we add the ARM Cortex M3 microcontroller integrated in the
GW1NSR-4C chip that powers the Sipeed TangNano 4K.

This primitive represents the microprocessor itself, but you need to add
ROM (flash), RAM, and some I/O to it for proper operation.

Luckily we have working flash, BSRAM and IOBUFs :)

The examples will show how to use this EMCU in the simplest cases.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>