Add EMCU blinky example
It shows the practical minimum of It shows the practical minimum of the
components that need to be added to the EMCU primitive to get a working
system.

You can use more memory addressing lines thereby increasing its volume
with the current 4K, as well as those who wish can make an APB bus to
communicate with other devices.

Firmware compilation is very difficult, but there is a good description:
https://github.com/verilog-indeed/gowin_fpga_tutorials/tree/main/gowin_empu

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>