EMCU. Add the APB example
The processor built into the GW1NSR-4C can communicate with the outside
world using another mechanism besides the Flash, memory and GPIO buses -
this is the Advanced Peripheral Bus (APB).

APB is part of the Advanced Microcontroller Bus Architecture hierarchy
of buses and is optimized for minimal power consumption and reduced
interface complexity.

APB should be used to interface to any peripherals which are
low-bandwidth and do not require the high performance of a piplined bus
interface.

Strictly speaking, according to the specification there must be a bridge
between the high-speed AHB bus and the APB bus, among the functions of
which are device selection (PSELx), generation of the PENABLE signal and
holding of values on the address bus.

And if PSELx signals have to be generated, there is no question, the
EMCU primitive already has a PENABLE output. So I took the liberty and
did not implement the latching of address lines, so the bridge turned
out to be very small:

``` verilog
  wire psel1;
  assign psel1 = apbtargexp2_psel && (apbtargexp2_paddr[11:8] == 4'h4); // 0x40002 >4< 00
```

Here I compare to the address of the first device - 0x40002400, there
are 12 such devices with base addresses spaced 256 bytes apart (this
does not cover the entire range of the 4 high bits of the APB address
bus, so there are probably some unknown devices with fixed addresses).

I don't use the [7:0] address bits in this example, but it's actually a
very useful thing to implement the registers of a particular peripheral.

The exchange is done in 32-bit words so 64 registers per device.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>