EMCU. Add a more complex APB device example.
Here we demonstrate the implementation of a more advanced peripheral to
the Tangnano4k embedded processor - the SPI controller as a device on
the APB.

It is very simple but nevertheless it is represented in the
address space of the processor as status, control and data registers.

The status register is used to indicate that the controller is busy, the
control register allows the bit order to be changed during transmission.

SPI is used to control the display, the picture is decompressed from ROM
and sent byte by by byte to the controller.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>