Add IEM (Input Edge Monitor)
This sensor will in the future allow dynamic control of the delay line
at the IBUF input. At the moment, the changes to the examples only
verify that the primitive compiles successfully.

This primitive is similar to MUXes in the sense that it has no fuses, it
is always on and it is up to the user to use the signals on its outputs
or not.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>