Implement MIPI and I3C primitives
Primitives for forming MIPI signals - MIPI_OBUF and MIPI_OBUF_A, as well
as primitive for receiving MIPI signals - MIPI_IBUF - are implemented.

These are not emulation, but real MIPI buffers, which are available in
chips GW1NR-9C and GW1NSR-4C (respectively Tangnano9k and Tangnano4k
boards).

The added examples are not use cases, but rather compilation checks.
This is due to the fact that the documentation for the actual protocol
of communication with MIPI devices is not open and manufacturers of
displays and cameras are very reluctant (actually never) to share the
description of registers and commands for their devices.

Also support of buffer primitive for exchange via I3C protocol is added
here. This protocol also belongs to the MIPI alliance and is not found
in open sources.

However, happy owners of documentation for their MIPI and I3C devices
can use these primitives.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>