Add segmented clock wires.
Briefly, these wires can be described as connecting directly to the
sinks in a rectangle 4 columns wide and at least 10 rows high. The
source can be either at the upper end or at the lower end, which is
connected by a system of ‘isolation’ of unused ends.

These small nets are useful when the global ones are exhausted or they
are not suitable for a given type of sink and/or source - these wires
are omnivorous.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>