Add BSRAM operation check
A femto-riscv-memtest example is added that contains riscv and firmware
to test memory writes/reads.

Memory is checked after the end of the program and almost up to the
maximum allowable address - whole chunks of 64 bytes are checked for
simplicity of the program itself.

Implemented for Tangnano4k, Tangnano9k and Tangnano20k with 16K, 48K and
88K BSRAM check respectively.

Firmware source code in examples/riscv-firmware/memtest-src directory.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>