Defining IO parameters in the examples
For all supported boards, we specify the IO standard (and resistors
where relevant) in the examples. We do this because the voltage
regulators on these boards are hard soldered and beginners can be guided
by these parameters of IO banks.

The exception is SZFPGA boards - there voltages can be easily changed by
jumpers and there is no sense to specify them here - those who bought
such boards are aware of what they do.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>