Fix VCC nets.
The new IDE has some peculiarities in the wire representation. In
particular fuses for VCC in chip description files are assigned to wires
Q6 and Q7, which is incomprehensible because DFF6 and DFF7 do not exist.

Further, the number of fuses itself is two less than in the old IDE.

Also there are ‘forgotten’ fuses at some inputs SEL7 and SEL3, but all
these things are taken into account and almost without losses bypassed.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>