Don't compress EMCU examples.
OpenFPGALoader frowns on checksum mismatch when loading images AND
firmware of the Tangnano4k embedded processor when using compression.

Since cpu firmware is compiled separately from the FPGA image and
completely different compilers and compression/uncompression
independently and, in addition, it is possible and in practice very
useful to use the same FPGA image, but change the CPU firmware, it is
clear that we can not specify the checksum in the FPGA image taking into
account the CPU firmware.

So I think that OpenFPGALoader's scolding here is groundless ;) and for
now we do not use image compression in the EMCU examples.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>