Put back the wires with partial fuses.
Revert back previously disabled suspicious PIPs that have less fuse in
the new IDE tables than in the old tables. As it turned out, the
“missing” fuzes are in the “alonenode” table, which is responsible for
setting fuzes in the situation when its sink is connected to none of
the listed sources.

Earlier there was no need in it because all the fuses were specified in
the main ['wire'][2] table.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>