GW5A. Implement simple IO and logic.
IO in this family turned out to be not so simple - blocks A and B can
(and in most cases will) be separated on different cells, respectively,
and wires (routing) and fuses must take this into account.

IO attributes defined experimentally are set, but their true nature is
left for the future.

This commit implements simple logic using LUT4. Higher order LUTs are
next in line.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>